move all course types to Course folder
